requirements.txt
av==6.1.0
Click==7.0
cycler==0.10.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
kiwisolver==1.0.1
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
pymediainfo==3.0
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.12.0
termcolor==1.1.0
